refactor: 콘솔 버전과 웹 버전의 컨트롤러 분리
