+++ src/controllers/WebController.js
@@ -1,0 +1,7 @@
+const WebController = {
+  start() {
+    console.log('start');
+  },
+};
+
+export default WebController;
